Reuse listing_results.html template for HistoryView results and specify no results message from the view
Not using index_results.html template as that one overrides the
no_results_message block with a message that's specific to the model,
which in this case doesn't apply because it's different from what the
listing model actually uses (the object's model vs ModelLogEntry).

The index_results.html template could probably be simplified to pass the
no_results_message from the view instead of defining it in the template,
but that requires refactoring of the is_searching and is_filtering
values into the view as well.