Update CircleCI's pipenv cache key to clear the cache
https://circleci.com/docs/caching/#clearing-cache

"Caches cannot be cleared. If you need to generate a new set of
caches you can update the cache key, similar to the previous
example. You might wish to do this if you have updated language
or build management tool versions."

"Caches are immutable, so it is helpful to start all your cache
keys with a version prefix, for example v1-.... This allows you
to regenerate all of your caches just by incrementing the
version in this prefix."