Ensure BaseDropdownMenuButton components render their attrs
Remove 'target' and 'rel' from the ButtonWithDropdownFromHook that is
used for the page listings' default "More" actions dropdown, as it does
not make sense to have those attributes.

These attributes were not actually rendered, because there's another bug
in the template: it used self.attrs, and 'self' does not exist in the
computed get_context_data(). Even if we add the dropdown instance as
'self', or reuse the 'button' variable from the BaseButton context,
there's still a problem: the {% dropdown %} tag expects it as a flat
string of attributes. This means we need to use flatatt() before using
it in the template.

Use the base_attrs_string property instead, which takes care of
flatatt() for us.