Use more descriptive values for GitHub Actions matrix config
GitHub only displays the values in the checks summary, so the values of
the `experimental` key only show up as "true" and "false", and
check_template_number_format as "1", which aren't obvious.

Use the strings "experimental" and "check-template" as the values
instead, and remove any experimental: false declarations. This will
make GitHub display 'experimental' instead of 'true', and only for
matrices that do have it set. It will also display 'check-template'
instead of '1'.